refactor: move Electron download store into its own platform domain
- Relocate src/stores/electronDownloadStore.ts to src/platform/electronDownload/
  so the store is colocated with its composable and UI components, matching the
  convention used by sibling domains (missingModel, missingMedia, assets, ...).
- Extract the Sentry failure reporter into a standalone
  src/platform/electronDownload/downloadFailureReporter.ts module so the store
  stays focused on IPC-backed state management and the reporter becomes
  straightforward to mock in tests.
- Update all eight import sites (sidebar, composables, missing-model download
  flow, and the test mock path). No behavior change.
