+++ src/platform/electronDownload/components/ElectronDownloadProgress.vue
@@ -70,7 +70,7 @@ import { useI18n } from 'vue-i18n'
 
 import Button from '@/components/ui/button/Button.vue'
 import { useElectronDownload } from '@/platform/electronDownload/composables/useElectronDownload'
-import type { ElectronDownload } from '@/stores/electronDownloadStore'
+import type { ElectronDownload } from '@/platform/electronDownload/electronDownloadStore'
 
 const { download } = defineProps<{
   download: ElectronDownload
